requirements.txt
Django==3.2.6
mysqlclient==2.0.3
django-cors-headers==3.8.0
djangorestframework==3.12.4
wheel==0.37.0
sendgrid==6.8.2